modif

+++ ft_strlen.c
@@ -14,9 +14,9 @@
 #include <string.h>
 #include <stddef.h>
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	length;
+	int	length;
 
 	length = 0;
 	while (s[length] != '\0')
